Allowed the in-memory context to fetch navigation properties
